[Button] Improve types with regard to react-router (#15193)
* [test] Extract Button type test into separate file

* [Button] Improve types with regard to react-router
